Extract maker variable drawer
